aws-serverless-logistics/functions/rastreamento/requirements.txt
boto3==1.34.0
geopy==2.4.0
haversine==2.8.0
PyJWT==2.8.0